mesh_gateway/requirements.txt
meshcore
paho-mqtt==1.6.1
pyserial
